requirements.txt
numpy<2
h5py>=3.10
Pillow>=10
PyYAML>=6
matplotlib>=3.7
torch>=2.1
torchvision>=0.16
tqdm>=4.66
